LU-3500 mdt: use __u64 for sp_cr_flags
In mdt_reconstruct_open() and mdt_open_by_fid(), when storing a copy
of the sp_cr_flags member of struct md_op_spec on the stack use __u64
rather than __u32. In mdd_create_data() use LPO64 for printing
sp_cr_flags.

Reviewed-on: http://review.whamcloud.com/6868
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/mdt/mdt_open.c
@@ -984,7 +984,7 @@ void mdt_reconstruct_open(struct mdt_thread_info *info,
         struct lsd_client_data  *lcd  = ted->ted_lcd;
         struct md_attr          *ma   = &info->mti_attr;
         struct mdt_reint_record *rr   = &info->mti_rr;
-        __u32                   flags = info->mti_spec.sp_cr_flags;
+	__u64                   flags = info->mti_spec.sp_cr_flags;
         struct ldlm_reply       *ldlm_rep;
         struct mdt_object       *parent;
         struct mdt_object       *child;
@@ -1083,11 +1083,10 @@ out:
 		lustre_msg_get_transno(req->rq_repmsg) == 0));
 }
 
-int mdt_open_by_fid(struct mdt_thread_info* info,
-                    struct ldlm_reply *rep)
+int mdt_open_by_fid(struct mdt_thread_info *info, struct ldlm_reply *rep)
 {
         const struct lu_env     *env = info->mti_env;
-        __u32                    flags = info->mti_spec.sp_cr_flags;
+	__u64			 flags = info->mti_spec.sp_cr_flags;
         struct mdt_reint_record *rr = &info->mti_rr;
         struct md_attr          *ma = &info->mti_attr;
         struct mdt_object       *o;
@@ -1127,7 +1126,7 @@ int mdt_open_by_fid_lock(struct mdt_thread_info *info, struct ldlm_reply *rep,
 {
         const struct lu_env     *env   = info->mti_env;
         struct mdt_device       *mdt   = info->mti_mdt;
-        __u32                    flags = info->mti_spec.sp_cr_flags;
+        __u64                    flags = info->mti_spec.sp_cr_flags;
         struct mdt_reint_record *rr    = &info->mti_rr;
         struct md_attr          *ma    = &info->mti_attr;
         struct mdt_object       *parent= NULL;
